some updates

+++ visualQC/graphicGenerator.py
@@ -10,12 +10,54 @@ import os
 import csv
 import re
 
+class NameModel():
+
+    """
+    Le model de nom d'un fichier :
+	prefix + infix + suffix + format
+
+    (si il ne devait y avoir, de façon certaine, qu'un station code dans le
+    graphique, ce code apparaitrait tel quel dans le nom de fichier à la
+    place du symbole "#S")
+
+    L'absence de #L et #C dans le nom de fichier indique que les informations
+    de location et channel codes ne sont pas présentes dans le graphique.
+
+    (concernant les informations qui peuvent être présentes dans les
+    différents types de graphique, l'ordre à respecter dans les modèles
+    de nom de fichier serait : NETCODE, STACODE, LOCCODE, CHACODE)
+
+    ex. 4G.#S.#L.AZBBA.TimeWaveformsC.jpeg
+    prefix = 4G.#S.#L.AZBBA
+    infix = TimeWaveformsC
+    suffix = ''
+    format = .jpeg
+
+    """
+
+    def __init__(self, prefix='', infix='', suffix=''):
+        self.prefix=prefix
+        self.infix=infix
+        self.suffix=suffix
+
+    def setPrefix(self, aPrefix):
+        self.prefix=aPrefix
+
+    def setInfix(self, anInfix):
+        self.infix=anInfix
+
+    def setSuffix(self, aSuffix):
+        self.suffix=aSuffix    
+
+    def generateName(self):
+        return self.prefix + self.infix + self.suffix
+
 class GraphicGenerator(ABC):
 
-    def __init__(self,inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None):
+    def __init__(self,inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None):
         self.inputDir=inputDir
         self.outputDir = outputDir
-        self.outSuffix = outSuffix
+        self.outputModel = outModel
         self.outputFile=outputFile
         self.outputFormat=outputFormat
 
@@ -29,8 +71,8 @@ class MetadataGraphicGenerator(GraphicGenerator):
     extension='.xml'
     iXMLFiles=[]
 
-    def __init__(self, inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None):
-        super().__init__(inputDir, outputDir, outSuffix, outputFile, outputFormat)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None):
+        super().__init__(inputDir, outputDir, outModel, outputFile, outputFormat)
         self.iMetaDataDir=inputDir
 
     def getInputFile(self, station=None):
@@ -63,8 +105,8 @@ class MetadataGraphicGenerator(GraphicGenerator):
 
 class TimeBasedGraphicGenerator(GraphicGenerator):
 
-    def __init__(self, inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None, startTime=None, endTime=None):
-        super().__init__(inputDir, outputDir, outSuffix,  outputFile, outputFormat)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None, startTime=None, endTime=None):
+        super().__init__(inputDir, outputDir, outModel,  outputFile, outputFormat)
         self.startTime=startTime
         self.endTime=endTime
 
@@ -74,8 +116,8 @@ class TimeBasedGraphicGenerator(GraphicGenerator):
 
 class EventBasedGraphicGenerator(TimeBasedGraphicGenerator):
 
-    def __init__(self, inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None, startTime=None, endTime=None, eventTime=None):
-        super().__init__(inputDir, outputDir, outSuffix, outputFile, outputFormat, startTime, endTime)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None, startTime=None, endTime=None, eventTime=None):
+        super().__init__(inputDir, outputDir, outModel, outputFile, outputFormat, startTime, endTime)
         self.eventTime=eventTime
 
     @abstractmethod
@@ -84,8 +126,8 @@ class EventBasedGraphicGenerator(TimeBasedGraphicGenerator):
 
 class DurationBasedGraphicGenerator(TimeBasedGraphicGenerator):
 
-    def __init__(self, inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None, startTime=None, endTime=None, duration=None):
-        super().__init__(inputDir, outputDir, outSuffix,  outputFile, outputFormat, startTime, endTime)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None, startTime=None, endTime=None, duration=None):
+        super().__init__(inputDir, outputDir, outModel,  outputFile, outputFormat, startTime, endTime)
         self.duration=duration
 
     @abstractmethod
@@ -99,8 +141,8 @@ class MeasuredDataGraphicGenerator(DurationBasedGraphicGenerator):
     loc="*"
     chan="*"
 
-    def __init__(self, inputDir, outputDir=None, outSuffix=None, outputFile=None, outputFormat=None, station="*", channel="*", startTime=None, endTime=None, duration=None):
-        super().__init__(inputDir, outputDir, outSuffix,  outputFile, outputFormat, startTime, endTime, duration)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None, station="*", channel="*", startTime=None, endTime=None, duration=None):
+        super().__init__(inputDir, outputDir, outModel,  outputFile, outputFormat, startTime, endTime, duration)
         self.iDataDir=inputDir
         self.sta=station
         self.chan=channel
@@ -134,12 +176,31 @@ class GraphicMetaData:
 
     def  generateCSV(self, data):
 
-        with open(self.csvFileName, 'w', newline='') as csvfile:
-            mywriter = csv.writer(csvfile, delimiter=',', quotechar='"', quoting=csv.QUOTE_NONNUMERIC)
-            mywriter.writerow(self.csvFieldNames)
-            mywriter.writerow(data)        
+        if os.path.exists(self.csvFileName):
+            with open(self.csvFileName, 'a') as csvfile:
+                mywriter = csv.writer(csvfile, delimiter=',', quotechar='"', quoting=csv.QUOTE_NONNUMERIC)
+                mywriter.writerow(data)
+        else: 
+            with open(self.csvFileName, 'w', newline='') as csvfile:
+                mywriter = csv.writer(csvfile, delimiter=',', quotechar='"', quoting=csv.QUOTE_NONNUMERIC)
+                mywriter.writerow(self.csvFieldNames)
+                mywriter.writerow(data) 
 
-class plotStationsMap(MetadataGraphicGenerator):
+class PlotStationsMap(MetadataGraphicGenerator):
+
+    def generateName(self):
+
+        # Olivier comments : 
+        #Le modèle de nom du fichier de sortie pourrait être, par défaut :
+        #"<NETCODE>.#S.StationsMap.jpeg"
+        # ou :
+        #"<NETCODE>.#S.stamap.jpeg"
+
+        outModel = self.outputModel
+        outFile = outModel.generateName() + self.outputFormat
+        if self.outputDir != None:
+            outFile = self.outputDir+outFile
+        return outFile
 
     def generate(self):
         """
@@ -153,15 +214,32 @@ class plotStationsMap(MetadataGraphicGenerator):
         #print(inv[0].code)
         #print(self.outSuffix)
         if  outFile == None:
-            outFile = inv[0].code+self.outSuffix
-            if self.outputDir != None:
-                outFile = self.outputDir+'/'+outFile
+            self.outputModel.setPrefix(inv[0].code+'.')
+            outFile = self.generateName()
         print('Generate image: '+outFile)
         inv.plot(projection="local", resolution="i", outfile=outFile)
 
-class plotDataAvailability(EventBasedGraphicGenerator):
+class PlotDataAvailability(EventBasedGraphicGenerator):
 
     command='obspy-scan'
+ 
+    def generateName(self):
+
+        # Olivier comments : 
+        #Le modèle de nom du fichier de sortie pourrait être, par défaut :
+        #"<NETCODE>.#S.#L.#C.DataAvailability.jpeg"
+        #ou :
+        #"<NETCODE>.#S.#L.#C.davail.jpeg"
+
+        #La présence de symboles "#S", "#L" et "#C" indique que station,
+        #location et channel codes sont présents, en tant qu'informations,
+        #dans le graphique.
+
+        outModel = self.outputModel
+        outFile = outModel.generateName() + self.outputFormat
+        if self.outputDir != None:
+            outFile = self.outputDir+outFile
+        return outFile  
 
     def generate(self):
         """
@@ -178,20 +256,31 @@ class plotDataAvailability(EventBasedGraphicGenerator):
 
         outFile = self.outputFile
         if  outFile == None:
-            outFile = self.outSuffix
-            if self.outputDir != None:
-                outFile = self.outputDir+'/'+outFile
+            outFile = self.generateName()
         self.command=self.command+' --output '+outFile
         print(self.command)
         os.system(self.command)
 
-class L2_nsplot_sInstrumentResponse(MetadataGraphicGenerator):
+class plotInstrumentResponseS(MetadataGraphicGenerator):
     
     station = ""
 
-    def __init__(self, inputDir,outputFile, outputFormat=None, station=""):
-        super().__init__(inputDir,outputFile, outputFormat)
+    def __init__(self, inputDir, outputDir=None, outputModel=None, outputFile=None, outputFormat=None, station=""):
+        super().__init__(inputDir, outputDir, outputModel, outputFile, outputFormat)
         self.station=station
+
+
+
+    def generateName(self):
+
+        #Le modèle de nom du fichier de sortie pourrait être, par défaut :
+        #"<NETCODE>.<STACODE>.#L.#C.InstrumentResponseS.jpeg"
+
+        outModel = self.outputModel
+        outFile = outModel.generateName() + self.outputFormat
+        if self.outputDir != None:
+            outFile = self.outputDir+outFile
+        return outFile
 
     def generate(self):
         """
@@ -199,9 +288,17 @@ class L2_nsplot_sInstrumentResponse(MetadataGraphicGenerator):
         """
 
         #inFile = self.getInputFile(self.station)
-        inFile =self.iMetaDataDir        
-        sta = read_inventory(inFile,'STATIONXML')[0][0]	
-        sta.plot(0.001, output="VEL", outfile=self.outputFile)
+        inFile =self.iMetaDataDir
+
+        inv = read_inventory(inFile,'STATIONXML')
+        outFile = self.outputFile
+        if  outFile == None:
+            self.outputModel.setPrefix(inv[0].code+'.'+self.station+'.')
+            outFile = self.generateName()
+                
+        sta = inv[0].select(station=self.station)[0]
+        print('Generate image: '+outFile)
+        sta.plot(0.001, output="VEL", outfile=outFile)
 
 class L3_nsplot_cInstrumentResponse(MetadataGraphicGenerator, GraphicMetaData):
 
@@ -222,11 +319,28 @@ class L3_nsplot_cInstrumentResponse(MetadataGraphicGenerator, GraphicMetaData):
         self.generateCSV([self.channel, self.outputFile])
 
 
-class  L2B_nsplot_sTimeWaveforms(MeasuredDataGraphicGenerator, GraphicMetaData):
+class  PlotTimeWaveformsS(MeasuredDataGraphicGenerator, GraphicMetaData):
 
-    def __init__(self, inputDir,outputFile, outputFormat=None, station="*", channel="*", startTime=None, endTime=None, duration=None, csvFileName="", csvFieldNames=[]):
-        MeasuredDataGraphicGenerator.__init__(self,inputDir,outputFile, outputFormat, station, channel, startTime, endTime, duration)
+    def __init__(self, inputDir, outputDir=None, outModel=None, outputFile=None, outputFormat=None, station="*", channel="*", startTime=None, endTime=None, duration=None, csvFileName="", csvFieldNames=[]):
+
+        MeasuredDataGraphicGenerator.__init__(self,inputDir, outputDir, outModel, outputFile, outputFormat, station, channel, startTime, endTime, duration)
         GraphicMetaData.__init__(self, csvFileName, csvFieldNames)
+        #self.outInfix = outInfix
+
+    def generateName(self):
+
+        # Olivier comments : Des informations temporelles pourraient être ajoutées :
+        # "<NETCODE>.<#STACODE>.<LOCCODE>.#C.<TIMERANGE>.STimeWaveforms.jpeg"
+        # sous la forme : "yyyymmddThhmmss-yyyymmddThhmmss"
+        # (année, mois et jour séparés, par le symbole "T", de heure, minute et seconde)
+
+        outModel = self.outputModel
+        outModel.setPrefix(self.sta+'.')
+        outFile = outModel.generateName() + self.outputFormat
+        #outFile = self.sta + self.outInfix + self.outFormat
+        if self.outputDir != None:
+            outFile = self.outputDir+outFile
+        return outFile
 
     def generate(self):
         """
@@ -238,9 +352,14 @@ class  L2B_nsplot_sTimeWaveforms(MeasuredDataGraphicGenerator, GraphicMetaData):
 
         st = self.getStream()
         print(st)
-        st.plot(outfile=self.outputFile, equal_scale = False)
+
+        outFile = self.outputFile
+        if  outFile == None:
+            outFile = self.generateName()
+
+        st.plot(outfile=outFile, equal_scale = False)
         #st.plot(outfile=self.outputFile)
-        self.generateCSV([self.sta, self.startTime, self.endTime, self.outputFile])
+        self.generateCSV([self.sta, self.startTime, self.endTime, outFile])
 
 class  L2B_nsplot_sTimeWaveforms_RemResp(MeasuredDataGraphicGenerator, MetadataGraphicGenerator, GraphicMetaData):
 
@@ -320,7 +439,7 @@ class L3B_nsplot_cTimeWaveforms_RemResp(MeasuredDataGraphicGenerator, MetadataGr
 
 
         st = self.getStream()
-        #print(st)
+        #print(st)csv/timeWaveformsS/
         st2 = st.copy()
         print("st2 : ", st2)
         plotFile="../Images/TWfComp/L3B_nsplot_cTWaveforms_RemResp_"+self.outUnit+"_stepPlot_"+self.sta+"_"+self.chan+".jpeg"
